+++ demo/workshop.js
@@ -2,6 +2,9 @@ export const workshop = {
   // Put your participants here!
   // Note: This key cannot be used inside templateData, because templateData gets flattened with `this`
   participants: ['Joren', 'Felix', 'Alex'],
+  admins: ['Joren'],
+  adminPassword: 'pineapples',
+  appKey: '(=]#bYS940q)T8S*dX1g;Sey)X3YhN|98B>4hwE:c2ew8QrN3);hQN?x"5#yUS',
   title: 'Frontend Workshop',
 
   // Put your data here to be used inside your scaffolding template files