+++ src/components/AppShell.js
@@ -4,21 +4,19 @@ import './AdminSidebar.js';
 import './loadAndSetDankMonoFont.js';
 import './ParticipantCapsule.js';
 import './SelectCookie.js';
-// Placeholder here, we transform this to resolve to the workshop.js
-// in the rootDir folder. This is a user-provided file
-// eslint-disable-next-line import/no-unresolved
-import { workshop } from './workshopImport.js';
+
+const setCustomCSSProps = () => {
+  document.body.style.setProperty('--cwk-color-primary', '#4e88c2');
+  document.body.style.setProperty('--cwk-color-secondary', '#34618e');
+  document.body.style.setProperty('--cwk-color-white', '#ffffff');
+  document.body.style.setProperty('--cwk-color-primary-transparent', '#4e88c230');
+};
 
 class AppShell extends LitElement {
   static get styles() {
     return css`
       :host {
         display: block;
-        --cwk-color-primary: #4e88c2;
-        --cwk-color-secondary: #34618e;
-        --cwk-color-white: #ffffff;
-        --cwk-color-primary-transparent: #4e88c230;
-
         height: 100vh;
         background-color: #fcfcfc;
       }
@@ -96,15 +94,30 @@ class AppShell extends LitElement {
     };
   }
 
+  constructor() {
+    super();
+    setCustomCSSProps();
+    this.fetchConfigComplete = new Promise(resolve => {
+      this.fetchConfigResolve = resolve;
+    });
+  }
+
   connectedCallback() {
     if (super.connectedCallback) {
       super.connectedCallback();
     }
-    this.getParticipants();
+    this.fetchWorkshopConfig();
     this.getParticipantName();
-    if (workshop.title) {
-      this.title = workshop.title;
-    }
+  }
+
+  async fetchWorkshopConfig() {
+    const { workshop } = await import(this.workshopImport || 'placeholder-import.js');
+    this.workshop = workshop;
+    const { title, participants } = this.workshop;
+    this.participants = participants;
+    this.title = title || this.title;
+    this.fetchConfigResolve();
+    return this.workshop;
   }
 
   getParticipantName() {
@@ -122,11 +135,6 @@ class AppShell extends LitElement {
     }
   }
 
-  async getParticipants() {
-    const { participants } = workshop;
-    this.participants = participants;
-  }
-
   changeName() {
     document.cookie = 'participant_name=;path=/;max-age=0';
     this.currentParticipantName = null;
@@ -137,7 +145,7 @@ class AppShell extends LitElement {
       ${this.currentParticipantName
         ? html`
             <div class="header">
-              <cwk-admin-sidebar></cwk-admin-sidebar>
+              <cwk-admin-sidebar .websocketPort=${this.websocketPort}></cwk-admin-sidebar>
               <button class="change-name" @click=${this.changeName}>Change your name</button>
               <p class="current-name">Hi ${this.currentParticipantName}!</p>
             </div>
@@ -146,13 +154,16 @@ class AppShell extends LitElement {
 
             <div>
               <div class="participants-container">
-                ${this.participants.map(
-                  name => html`<cwk-participant-capsule .name="${name}"></cwk-participant-capsule>`,
-                )}
+                ${this.participants
+                  ? this.participants.map(
+                      name =>
+                        html`<cwk-participant-capsule .name="${name}"></cwk-participant-capsule>`,
+                    )
+                  : html``}
               </div>
             </div>
           `
-        : html`<cwk-select-cookie></cwk-select-cookie>`}
+        : html`<cwk-select-cookie .workshopImport=${this.workshopImport}></cwk-select-cookie>`}
     `;
   }
 }