+++ src/middlewares/middlewares.js
@@ -1,19 +1,5 @@
-import { adminUIMiddleware } from './admin-ui-middleware.js';
 import { changeParticipantUrlMiddleware } from './change-participant-url-middleware.js';
-import { createFileControlMiddleware } from './file-control-middleware.js';
-import { createInsertAppShellMiddleware } from './insert-app-shell-middleware.js';
-import { insertFollowModeScriptMiddleware } from './insert-follow-mode-script-middleware.js';
+import { jwtMiddleware } from './jwt-middleware.js';
 import { noCacheMiddleware } from './no-cache-middleware.js';
-import { createWorkshopImportReplaceMiddleware } from './workshop-import-replace-middleware.js';
-import { createWsPortReplaceMiddleware } from './ws-port-replace-middleware.js';
 
-export {
-  createFileControlMiddleware,
-  createInsertAppShellMiddleware,
-  noCacheMiddleware,
-  createWorkshopImportReplaceMiddleware,
-  adminUIMiddleware,
-  insertFollowModeScriptMiddleware,
-  changeParticipantUrlMiddleware,
-  createWsPortReplaceMiddleware,
-};
+export { noCacheMiddleware, changeParticipantUrlMiddleware, jwtMiddleware };