+++ src/start-server.js
@@ -1,22 +1,20 @@
 import commandLineArgs from 'command-line-args';
-import {
-  commandLineOptions,
-  createConfig,
-  readCommandLineArgs,
-  startServer as startEdsServer,
-} from 'es-dev-server';
+import { commandLineOptions, createConfig, startServer as startEdsServer } from 'es-dev-server';
 import path from 'path';
 import WebSocket from 'ws';
 import {
-  adminUIMiddleware,
   changeParticipantUrlMiddleware,
-  createFileControlMiddleware,
-  createInsertAppShellMiddleware,
-  createWorkshopImportReplaceMiddleware,
-  createWsPortReplaceMiddleware,
-  insertFollowModeScriptMiddleware,
+  jwtMiddleware,
   noCacheMiddleware,
 } from './middlewares/middlewares.js';
+import {
+  adminUIPlugin,
+  appShellPlugin,
+  fileControlPlugin,
+  followModePlugin,
+  workshopImportPlugin,
+  wsPortPlugin,
+} from './plugins/plugins.js';
 import { cwkState } from './utils/CwkStateSingleton.js';
 
 const getAdminUIDefaults = () => {
@@ -37,6 +35,10 @@ const sendAdminConfig = ws => {
   );
 };
 
+const setDefaultAdminConfig = () => {
+  cwkState.state = { adminConfig: getAdminUIDefaults() };
+};
+
 const handleWsMessage = (message, ws) => {
   const parsedMessage = JSON.parse(message);
   const { type } = parsedMessage;
@@ -46,9 +48,23 @@ const handleWsMessage = (message, ws) => {
       sendAdminConfig(ws);
       break;
     }
+    case 'reset-state': {
+      setDefaultAdminConfig();
+      break;
+    }
+    case 'clear-state': {
+      cwkState.clear();
+      break;
+    }
     case 'config-updated': {
       const { config } = parsedMessage;
       cwkState.state = { adminConfig: config };
+      ws.send(
+        JSON.stringify({
+          type: 'config-update-completed',
+          config: cwkState.state.adminConfig,
+        }),
+      );
       break;
     }
     case 'authenticate': {
@@ -61,6 +77,12 @@ const handleWsMessage = (message, ws) => {
         // Store the websocket connection for this user
         state.wsConnections.set(username, ws);
         cwkState.state = state;
+        ws.send(
+          JSON.stringify({
+            type: 'authenticate-completed',
+            user: username,
+          }),
+        );
       }
     }
     // no default
@@ -74,18 +96,52 @@ const setupWebSocket = wsPort => {
     ws.on('message', message => handleWsMessage(message, ws));
   });
 
-  cwkState.state = { wss };
   return wss;
 };
 
-export const startServer = async (opts = {}) => {
+const addPluginsAndMiddlewares = (edsConfig, cwkConfig) => {
+  const newEdsConfig = edsConfig;
+  newEdsConfig.plugins = [...edsConfig.plugins];
+  newEdsConfig.middlewares = [...edsConfig.middlewares];
+
+  /**
+   * Right now, we assume that your workshop.js is in the same folder as your app index.
+   * TODO: allow override.
+   */
+  const absoluteRootDir = path.resolve('/', path.dirname(cwkConfig.appIndex));
+
+  newEdsConfig.plugins.push(wsPortPlugin(cwkConfig.wsPort));
+  newEdsConfig.plugins.push(workshopImportPlugin(absoluteRootDir));
+  newEdsConfig.plugins.push(followModePlugin(absoluteRootDir, cwkConfig.wsPort));
+  newEdsConfig.middlewares.push(changeParticipantUrlMiddleware(absoluteRootDir));
+  newEdsConfig.middlewares.push(jwtMiddleware(absoluteRootDir));
+
+  // Plugins & middlewares that can be turned off completely from the start through cwk flags
+  if (!cwkConfig.alwaysServeFiles) {
+    newEdsConfig.plugins.push(
+      fileControlPlugin({ exts: ['js', 'html'], rootDir: absoluteRootDir }),
+    );
+  }
+
+  if (!cwkConfig.withoutAppShell) {
+    newEdsConfig.plugins.push(appShellPlugin(cwkConfig.appIndex, cwkConfig.title));
+    newEdsConfig.plugins.push(adminUIPlugin(absoluteRootDir));
+  }
+
+  if (!cwkConfig.enableCaching) {
+    newEdsConfig.middlewares.push(noCacheMiddleware);
+  }
+
+  return newEdsConfig;
+};
+
+const getCwkConfig = opts => {
   // cwk defaults
   let cwkConfig = {
     withoutAppShell: false,
     enableCaching: false,
     alwaysServeFiles: false,
     appIndex: './index.html',
-    port: 8000,
     wsPort: 8001,
     title: '',
     ...opts,
@@ -109,9 +165,9 @@ export const startServer = async (opts = {}) => {
         name: 'enable-caching',
         type: Boolean,
         description: `
-          If set, re-enable caching. By default it is turned off, since it's more often a hassle than a help in a workshop dev server
-          This also means that the file control middleware only has effect the upon first load, because the server serves cached responses.
-        `,
+        If set, re-enable caching. By default it is turned off, since it's more often a hassle than a help in a workshop dev server
+        This also means that the file control middleware only has effect the upon first load, because the server serves cached responses.
+      `,
       },
       {
         name: 'always-serve-files',
@@ -129,72 +185,54 @@ export const startServer = async (opts = {}) => {
     cwkConfig = {
       ...cwkConfig,
       ...commandLineArgs(cwkServerDefinitions, { argv: opts.argv }),
-      ...readCommandLineArgs(opts.argv),
     };
 
-    // TODO: reuse logic that eds readCommandLineUses to camelCase the cwk flags instead of syncing them here
+    // TODO: auto camelCase these kebab cased flags
     cwkConfig.withoutAppShell = cwkConfig['without-app-shell'] || cwkConfig.withoutAppShell;
     cwkConfig.enableCaching = cwkConfig['enable-caching'] || cwkConfig.enableCaching;
     cwkConfig.alwaysServeFiles = cwkConfig['always-serve-files'] || cwkConfig.alwaysServeFiles;
     cwkConfig.wsPort = cwkConfig['ws-port'] || cwkConfig.wsPort;
+    cwkConfig.appIndex = cwkConfig['app-index'] || cwkConfig.appIndex;
   }
 
-  /**
-   * Right now, we assume that your workshop.js is in the same folder as your app index.
-   * TODO: allow override.
-   */
-  const absoluteRootDir = path.resolve('/', path.dirname(cwkConfig.appIndex));
+  return cwkConfig;
+};
 
+const getEdsConfig = (opts, cwkConfig) => {
   // eds defaults & middlewares
-  const edsConfig = {
-    open: true,
+  let edsConfig = {
+    open: false,
+    logStartup: true,
     watch: false,
     moduleDirs: ['node_modules'],
     nodeResolve: true,
     logErrorsToBrowser: true,
-    middlewares: [
-      insertFollowModeScriptMiddleware,
-      ...(cwkConfig.withoutAppShell
-        ? []
-        : [createInsertAppShellMiddleware(cwkConfig.appIndex, cwkConfig.title)]),
-      ...(cwkConfig.enableCaching ? [] : [noCacheMiddleware]),
-
-      ...(cwkConfig.alwaysServeFiles
-        ? []
-        : [
-            createFileControlMiddleware({
-              ext: 'js',
-              admin: true,
-              rootDir: absoluteRootDir,
-            }),
-            createFileControlMiddleware({
-              ext: 'html',
-              admin: true,
-              rootDir: absoluteRootDir,
-            }),
-          ]),
-
-      createWorkshopImportReplaceMiddleware(absoluteRootDir),
-      createWsPortReplaceMiddleware(cwkConfig.wsPort),
-      adminUIMiddleware,
-      changeParticipantUrlMiddleware,
-    ],
+    compatibility: 'none',
+    plugins: [],
+    middlewares: [],
+    ...opts,
   };
 
-  const config = createConfig({
-    ...edsConfig,
-    ...cwkConfig,
-  });
+  edsConfig = addPluginsAndMiddlewares(edsConfig, cwkConfig);
+  edsConfig = createConfig(edsConfig);
+  return edsConfig;
+};
 
-  await startEdsServer(config);
+export const startServer = async (opts = {}) => {
+  const cwkConfig = getCwkConfig(opts);
+  const edsConfig = getEdsConfig(opts, cwkConfig);
+
+  const { server } = await startEdsServer(edsConfig);
   const wss = setupWebSocket(cwkConfig.wsPort);
 
-  cwkState.state = { adminConfig: getAdminUIDefaults() };
+  cwkState.state = { wss };
+  setDefaultAdminConfig();
 
   ['exit', 'SIGINT'].forEach(event => {
     process.on(event, () => {
       wss.close();
-      process.exit(0);
     });
   });
+
+  return { server, edsConfig, cwkConfig, wss };
 };