+++ test/utils/workshop-mock/template/main.js
@@ -1,0 +1,2 @@
+console.log(`Hi <%= participantNameLowercase %>, welcome to <%= appTitle %>!`);
+console.log(`<%= intro %>`);