+++ web-test-runner.config.js
@@ -1,0 +1,7 @@
+const {
+  puppeteerLauncher,
+} = require('web-test-runner/dist/implementations/browser-launchers/puppeteer-launcher.js');
+
+module.exports = {
+  browsers: [puppeteerLauncher()],
+};